+++ server/lib/nl/fulfillment/comparison.py
@@ -15,13 +15,13 @@
 import logging
 from typing import List
 
-from server.lib.nl.detection import Place
+from server.lib.nl.detection.types import Place
 from server.lib.nl.fulfillment.base import add_chart_to_utterance
-from server.lib.nl.fulfillment.base import ChartVars
 from server.lib.nl.fulfillment.base import populate_charts_for_places
-from server.lib.nl.fulfillment.base import PopulateState
 from server.lib.nl.fulfillment.context import \
     places_for_comparison_from_context
+from server.lib.nl.fulfillment.types import ChartVars
+from server.lib.nl.fulfillment.types import PopulateState
 from server.lib.nl.utterance import ChartOriginType
 from server.lib.nl.utterance import ChartType
 from server.lib.nl.utterance import Utterance
