+++ server/lib/nl/fulfillment/size_across_entities.py
@@ -16,14 +16,14 @@ import logging
 from typing import List
 
 from server.lib.nl import utils
-from server.lib.nl.detection import ContainedInPlaceType
-from server.lib.nl.detection import Place
-from server.lib.nl.detection import RankingType
-from server.lib.nl.detection import SizeType
+from server.lib.nl.detection.types import ContainedInPlaceType
+from server.lib.nl.detection.types import Place
+from server.lib.nl.detection.types import RankingType
+from server.lib.nl.detection.types import SizeType
 from server.lib.nl.fulfillment.base import add_chart_to_utterance
-from server.lib.nl.fulfillment.base import ChartVars
 from server.lib.nl.fulfillment.base import populate_charts
-from server.lib.nl.fulfillment.base import PopulateState
+from server.lib.nl.fulfillment.types import ChartVars
+from server.lib.nl.fulfillment.types import PopulateState
 from server.lib.nl.utterance import ChartOriginType
 from server.lib.nl.utterance import ChartType
 from server.lib.nl.utterance import Utterance
